move shared type to perseus-core
