+++ packages/perseus-core/src/index.ts
@@ -1,1 +1,2 @@
 export type {PerseusAnalyticsEvent, SendEventFn} from "./analytics";
+export type {KEScore, RendererInterface} from "./types";
