+++ packages/perseus/src/index.ts
@@ -108,11 +108,9 @@ export type {
     ImageUploader,
     JiptLabelStore,
     JiptRenderer,
-    KEScore,
     LinterContextProps,
     PerseusDependencies,
     PerseusScore,
-    RendererInterface,
     Version,
     VideoData,
     VideoKind,
