+++ testing/item-renderer-with-debug-ui.tsx
@@ -9,7 +9,8 @@ import KEScoreUI from "./ke-score-ui";
 import SideBySide from "./side-by-side";
 
 import type {PerseusItem} from "../packages/perseus/src/perseus-types";
-import type {APIOptions, KEScore} from "../packages/perseus/src/types";
+import type {APIOptions} from "../packages/perseus/src/types";
+import type {KEScore} from "@khanacademy/perseus-core";
 
 type Props = {
     item: PerseusItem;
