+++ testing/ke-score-ui.tsx
@@ -2,7 +2,7 @@ import {HeadingSmall} from "@khanacademy/wonder-blocks-typography";
 import * as React from "react";
 import ReactJson from "react-json-view";
 
-import type {KEScore} from "../packages/perseus/src/types";
+import type {KEScore} from "@khanacademy/perseus-core";
 
 type Props = {
     score: KEScore | null | undefined;
