+++ testing/multi-item-renderer-with-debug-ui.tsx
@@ -9,7 +9,8 @@ import KEScoreUI from "./ke-score-ui";
 import SideBySide from "./side-by-side";
 
 import type {Item as MultiItem} from "../packages/perseus/src/multi-items/item-types";
-import type {APIOptions, KEScore} from "../packages/perseus/src/types";
+import type {APIOptions} from "../packages/perseus/src/types";
+import type {KEScore} from "@khanacademy/perseus-core";
 
 type SimpleItemRenderTree = {
     blurb: React.ReactNode;
